+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 3.8, 2026-09-01
 ---------------
 - improve hero parsing speed (issue #32)
+- remember hero index sort order (issue #32)
 
 
 3.7, 2026-07-31